NeighborlyAPI/requirements.txt
azure-functions==1.11.2
pymongo==4.2.0
